refactor: add pageMeta to CmsContent type and rename function to buildCmsContentAndTranslations
